requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
frozenlist==1.4.1
geographiclib==2.0
idna==3.6
markdown-it-py==3.0.0
mdurl==0.1.2
meraki==1.45.0
multidict==6.0.5
pillow==10.3.0
Pygments==2.17.2
python-dotenv==1.0.1
requests==2.31.0
rich==13.7.1
urllib3==2.2.1
yarl==1.9.4
